fix(display): reserve prompt+input height at window bottom (#98)
When a multi-line primary prompt is rendered at the bottom of the terminal
window, the input area's trailing row does not exist, so the terminal clamps
the engine's downward cursor moves while the paired upward moves still travel.
The row bookkeeping (startRows/cursorRow) drifts and the prompt's lines get
overwritten -- the prompt "overlaps with the input line" reported in #98 and
reeflective/console#78.

ensureInputSpace() previously reserved space only for the input rows and bailed
out entirely for single-row input (lineRows <= 1), so a multi-line prompt with
a normal input line at the bottom got no reservation at all. It now reserves
lineRows+1 rows (the input rows plus one trailing row) below startRows and
scrolls the screen up by the deficit, computed from the already-probed
startRows so no extra ESC[6n query is issued.

Tests move under internal/display (external display_test package, so they can
drive the full shell without an import cycle) with issue-neutral names:
- TestRenderPromptAndInput, TestRenderWrappingAlignment
- TestRenderMultilinePromptAtBottom: regression guard, passes with the fix
- TestRenderMisreportedCursor: a separate cursor-probe robustness gap, gated
  behind READLINE_RUN_KNOWN_GAPS so CI stays green

Known remaining (pre-existing) edge case: a multi-line prompt AND a heavily
wrapping input both at the very bottom of a small window can still scroll the
topmost prompt line off, because upper prompt lines are printed once and not
repainted.
